refactor(monorepo): use ryunix-doc workspace instead of test/webpack copies
chore(scripts): replace setup-test-webpack with setup-ryunix-doc
docs(guide): document ../ryunix-doc integration and gitignored test/*
