+++ docs/es/guias/guia-del-repositorio.md
@@ -227,7 +227,7 @@ Cerebro del workspace: scripts globales y devDependencies compartidas.
 | `pnpm test`                              | Tests.                                                  |
 | `pnpm lint`                              | ESLint en raíz + lint por paquete.                      |
 | `pnpm release:canary` / `release:stable` | Versionado (`gmvu`) y publicación.                      |
-| `pnpm run:web`                           | App de prueba en `test/webpack` (si existe localmente). |
+| `pnpm run dev:doc`                       | Sitio de docs en `../ryunix-doc` (workspace).           |
 
 El repo raíz es `private: true` y no se publica a npm; sí lo hacen los paquetes
 en `packages/`.
@@ -239,7 +239,8 @@ Define los workspaces de pnpm:
 ```yaml
 packages:
   - 'packages/*'
-  - 'test/*'
+  - '_ci/smoke-app'
+  - '../ryunix-doc'
 ```
 
 #### `turbo.json`
@@ -343,17 +344,16 @@ navegador**.
 ```text
 
 1. pnpm install
-2. Crear app local en test/webpack (una vez; ver app de integración)
-3. pnpm build
-4. pnpm run:web          ← aquí desarrollas como en una app normal
-5. Si tocas packages/core → pnpm --filter @unsetsoft/ryunixjs build → refrescar navegador
-6. Si tocas packages/ryunix-presets → reiniciar pnpm run:web
+2. pnpm run setup:web   (una vez; ver app de integración)
+3. pnpm run dev:doc     ← desarrollo en el navegador
+4. Si tocas packages/core → pnpm --filter @unsetsoft/ryunixjs build → F5
+5. Si tocas packages/ryunix-presets → reiniciar pnpm run dev:doc
 ```
 
-**Analogía:** React/Next se desarrollan con una **app de ejemplo** o el propio
-sitio de docs; Ryunix usa `test/webpack` (gitignored, cada uno la crea en su
-máquina). Los scripts `run:web`, `run:web:build` y `run:web:start` son atajos
-desde la raíz hacia esa app.
+**Analogía:** React/Next se desarrollan con el propio sitio de docs; Ryunix usa
+el repo hermano **`ryunix-doc`** enlazado con `workspace:*`. `run:web` es alias
+de `dev:doc`. No copies el sitio bajo `test/webpack*`: esa carpeta está en
+`.gitignore` y no se versiona.
 
 Detalle por paquete (tests, CRA, DevTools):
 [tests automatizados](./tests-automatizados.md),
@@ -365,7 +365,7 @@ Detalle por paquete (tests, CRA, DevTools):
 
 ```bash
 pnpm install      # instalar dependencias de todo el monorepo
-pnpm run:web      # ← desarrollo en navegador (app en test/webpack)
+pnpm run dev:doc    # ← desarrollo en navegador (../ryunix-doc)
 pnpm build        # compilar paquetes (necesario antes de run:web la primera vez)
 pnpm test         # tests automatizados (solo core, Jest)
 pnpm lint         # lint global
