+++ docs/es/guias/pila-tecnologica-y-scripts.md
@@ -33,7 +33,7 @@ Referencia de **tecnologías usadas en el monorepo RyunixJS** y de los **scripts
 | :----------------------------------------------------- | :-------------------------------------------------------- |
 | Lenguajes, runtimes y bibliotecas principales por área | Tutoriales paso a paso de cada herramienta                |
 | Scripts del `package.json` raíz y tareas Turbo         | Cada flag de `ryunix` o Webpack                           |
-| Scripts por paquete a nivel resumen                    | Apps generadas en `test/webpack` (local, en `.gitignore`) |
+| Scripts por paquete a nivel resumen                    | `../ryunix-doc` (workspace) y pruebas locales en `test/*` |
 
 Para onboarding y estructura de carpetas, ver
 [guia-del-repositorio.md](./guia-del-repositorio.md). Para tests y
@@ -153,7 +153,7 @@ Los comandos se ejecutan desde la **raíz del repositorio** salvo indicación.
 
 | Script                   | Propósito                                                      |
 | :----------------------- | :------------------------------------------------------------- |
-| `pnpm run run:web`       | `ryunix dev` sobre `test/webpack` (gitignored; crear en local) |
+| `pnpm run run:web`       | Alias de `dev:doc` (`ryunix dev` en `../ryunix-doc`)           |
 | `pnpm run run:web:build` | Build de producción de esa app                                 |
 | `pnpm run run:web:start` | `ryunix start` de la app compilada                             |
