+++ pnpm-workspace.yaml
@@ -1,7 +1,5 @@
 packages:
   - 'packages/*'
-  - 'test/*'
-  - 'test/webpack/*'
   - '_ci/smoke-app'
-  # Sibling docs site (ryx/ryunix-doc) — use workspace:* in its package.json
+  # Canonical docs site (sibling repo). Not copied under test/webpack*.
   - '../ryunix-doc'
